requirements.txt
Werkzeug
tk
flask_mail
email_validator
Pillow
Flask==2.3.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Session==0.5.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.1
redis==5.0.3
pymongo==4.6.1
itsdangerous==2.1.2
authlib==1.2.1
google-generativeai==0.4.1
wtforms==3.1.2
MarkupSafe==2.1.5
